pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ibef-backend"
version = "0.1.0"
description = "Backend API for IBEF built with FastAPI"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "IBEF Team", email = "dev@example.com" },
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic-settings>=2.5.2",
]
optional-dependencies = { dev = [
  "httpx>=0.27.2",
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.6.0",
  "mkdocstrings[python]>=0.26.0",
  "mypy>=1.11.1",
  "pytest>=8.3.2",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.9",
] }

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.hatch.envs.default]
dependencies = ["ibef-backend[dev]"]

[tool.hatch.envs.default.scripts]
api = "cd src && uvicorn main:app --reload --host 0.0.0.0 --port 8000"
test = "pytest"
lint = "ruff check src tests"
typecheck = "mypy src"

[tool.hatch.envs.dev]
dependencies = ["ibef-backend[dev]"]

[tool.hatch.envs.docs]
dependencies = [
  "mkdocs>=1.6.0",
  "mkdocs-material>=9.6.0",
  "mkdocstrings[python]>=0.26.0",
]

[tool.hatch.envs.docs.scripts]
serve = "mkdocs serve"
build = "mkdocs build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
pythonpath = ["src"]
norecursedirs = ["src", "storage", "docs", ".git", ".hatch"]
